+++ compiler-tests/src/test/data/dump/ir/dependencygraph/InlinedObjectProviderUsesGeneratedFactory.kt
@@ -1,0 +1,11 @@
+object InlinedObject
+
+@BindingContainer
+object Bindings {
+  @Provides fun provideObject(): InlinedObject = InlinedObject
+}
+
+@DependencyGraph(bindingContainers = [Bindings::class])
+interface AppGraph {
+  val objectProvider: Provider<InlinedObject>
+}